requirements.txt
anyio==3.6.2
asyncclick==8.1.3.4
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
importlib-metadata==6.1.0
pydantic==1.10.7
python-kasa==0.5.1
requests==2.28.2
sniffio==1.3.0
typing_extensions==4.5.0
urllib3==1.26.15
zipp==3.15.0
